Extract product drives card to separate report [#3734]
