+++ app/controllers/dashboard_controller.rb
@@ -5,8 +5,6 @@ class DashboardController < ApplicationController
   def index
     setup_date_range_picker
 
-    @donations = current_organization.donations.during(helpers.selected_range)
-    @recent_donations = @donations.recent
 
     distributions = current_organization.distributions.includes(:partner).during(helpers.selected_range)
     @itemized_distribution_data = DistributionItemizedBreakdownService.new(organization: current_organization, distribution_ids: distributions.pluck(:id)).fetch
