+++ app/controllers/reports/product_drives_controller.rb
@@ -1,0 +1,7 @@
+class Reports::ProductDrivesController < ApplicationController
+  def index
+    setup_date_range_picker
+    @donations = current_organization.donations.during(helpers.selected_range)
+    @recent_donations = @donations.recent
+  end
+end
