+++ app/views/reports/product_drives/_product_drive.html.erb
@@ -1,0 +1,5 @@
+<div class="float-center donation">
+  <%= link_to donation do %>
+    <%= number_with_delimiter donation.line_items.total %> from <%= donation.product_drive&.name || donation.product_drive_participant.business_name %>
+  <% end %>
+</div>
